Fixed steamvr changind physics update
For some reason steamvr has a rule set to true by default that changes the rate of physics update, making the fixed size timeBody unusable.
Two fixes are possible in my case :
- Go back to using a list instead of an array
- Set the rule to false
I chose the second one because it seems simpler and I'm not a fan of changing mid game the physics update rate but who knows maybe it'll break something later on we'll see.