Added a storageArray that remove garbage collection
I was getting freezes every now and then because of garbage collection since I was replacing the pointsInTime array by another and leaving the old one to the garbage collector.
Now I use a second array that is only used to store/replace pointsInTime.
Also I used the profiler which is confusing but pretty cool I guess.