Remove line that blocked physics updates in StartXR which made hands jittery
Turns out I was right the rule was indeed sus.
It caused an error but was actually still done but only after updates even tho I set it to false in awake what the actual fuck.
So basically I either leave it like that and it changes the physics call only once which isn't so bad for vr but then the hands are jittery.
Or I remove it and then the recording of time isn't in seconds in it's seconds but sorta not idk it's weird. Also even when I remove it the hands are kinda jittery so is it really worth it ???(probably is).
Or I find a way to use coroutines assuming they work as they should.

Just saw that jittering was less visible in other scene so fair to say it's a performance problem more than anything.
So basically it all comes back to timeBody is unoptimized.
pain