Managed to get custom hands working kinda
Imported the default hand into blender -> sculpted on the mesh a little, export with unity FBX add on, das it.
Turns out whatever thing I did that moved the bones (i forgor :skull:) was probably fucking everything up.

Still need to test with entirely new mesh and also I'm not sure if I have to make the hands completely separate from body when modeling or if there's a way to cheese it.